feat(cypress-mock-framework): split into separate directories
