+++ tsconfig.json
@@ -23,5 +23,5 @@
     "allowUnreachableCode": false,
     "pretty": true
   },
-  "include": ["src/lib/testing/cypress/*.ts", "./node_modules/cypress"]
+  "include": ["src/lib/testing/cypress/**/*.ts", "./node_modules/cypress"]
 }
